refactor: extract cli, error, and utilities services from MyFlaskApp
